第八次上传： 1.修改Object的坐标数据类型，实现怪物的连贯移动 2.增加了Bullet类，并在Tower、Chapter中添加了相关方法。现在防御塔可以发生子弹攻击。

+++ Tower_Defence/Object.cpp
@@ -17,7 +17,7 @@ void Object::show(QPainter *painter)
 
 double dist(Object p1, Object p2)
 {
-    int xx = p1.getX() - p2.getX();
-    int yy = p1.getY() - p2.getY();
+    double xx = p1.getX() - p2.getX();
+    double yy = p1.getY() - p2.getY();
     return sqrt(xx*xx + yy*yy);
 }
