+++ Tower_Defence/Object.h
@@ -12,17 +12,17 @@ class Object
     public:
         Object(string type);
         void show(QPainter *painter);
-        void setX(int x) {this->_x = x;};
-        void setY(int y) {this->_y = y;};
-        int getX() const {return this->_x;};
-        int getY() const {return this->_y;};
+        void setX(double x) {this->_x = x;};
+        void setY(double y) {this->_y = y;};
+        double getX() const {return this->_x;};
+        double getY() const {return this->_y;};
         QImage getPic() const {return this->_pic;};
-        int getHeight() const {return this->_ic.getHeight();};
-        int getWidth() const {return this->_ic.getWidth();};
+        double getHeight() const {return this->_ic.getHeight();};
+        double getWidth() const {return this->_ic.getWidth();};
         string getType() const {return this->_ic.getType();};
         friend double dist(Object p1, Object p2);
     protected:
-        int _x, _y;//在画面中的位置
+        double _x, _y;//在画面中的位置
         QImage _pic;//图片
         Icon _ic;
 };
